pst-analysis-engine/requirements.txt
# Combined requirements from VeriCase Analysis and vericase-docs-rapid-plus-ts

# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
SQLAlchemy==2.0.35
alembic==1.13.1
psycopg2-binary==2.9.9  # For PostgreSQL

# Pydantic and Config
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.1.0

# AWS & Cloud Services
boto3==1.35.20
opensearch-py==2.6.0
requests-aws4auth==1.2.3

# Celery and Redis (for background tasks)
redis==5.0.8
celery==5.4.0

# File Handling & Processing
python-multipart==0.0.9
aiofiles==24.1.0
requests==2.32.3
PyPDF2==3.0.1          # PDF manipulation for watermarking etc.
reportlab==4.1.0        # PDF creation
python-magic==0.4.27    # File type detection

# AI / ML Models (from advanced build)
sentence-transformers==3.3.0
transformers==4.46.0
torch==2.5.1

# General Utilities
numpy==1.26.4 # Kept higher version
pandas==2.1.4
pyyaml==6.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
